block-jobs and dirty-bitmaps patches

# -----BEGIN PGP SIGNATURE-----
#
# iQIzBAABCgAdFiEEi5wmzbL9FHyIDoahVh8kwfGfefsFAmqDKl4ACgkQVh8kwfGf
# efsQ6A//SbHOBtTGr5h541r/ag55slCpSSbIZ92HxNkXAtFOiiPwMj2p+gkKH96m
# 9AMZslC8WFlrY4oAE9jWe3gnfrV43b+DlpvgeTocJjPMscLvt0BeHWgWrFlu6HkR
# RrXpHM7h75zHD16qON0ZxLWMR/VtN4JJNo0Q8VkVt7GvxrW0QNMfH9LlGHsGwnOS
# 3o5BW0BhtIFb4o+rMa5RY43FdghumXVjHZR84iJUvyhSrGiGjq+lZn5a+PhggO1b
# G3IR+/ZcpZv+FKNN0vRbmmgSOFmz0QR5LxQxERyyFajpfhNbUzXae2IbjM+VThSU
# G8JRGC8iL3js3qQdw8JsDEbegsfJniNoKoxolmQx8wAfmfEcHhAnSWwLsKjJttj4
# FVwrF7Sa33xdkztlDdAjAfUANm4u6q3zhy4Ap3mm4t45c9u0tNYtKz/jySgmgoW+
# YGSSWrP0eyXahe7J+BP+TkKooG2sb3mxczOUDhpQoySwgOw/jR4s4CvIn7Cs10WI
# hqHbzZJ5wCfIQK/7E+MfBi0W47amlMdwZ5B42pvAoiBnWsXA2vjl8JTtmI2H25Ax
# bPM0iNOkx1aJK+RMhWW++M1z+jXR246iQHipbpcQsyfFgB8mqKAytbvZ+uXp55jQ
# e8/D8FLVx8zwAVMl7rLFYxo4ntNyTeAfq2akJM5DZd0g7BWvfIM=
# =iQVf
# -----END PGP SIGNATURE-----
# gpg: Signature made Mon 17 Aug 2026 08:35:58 AM PDT
# gpg:                using RSA key 8B9C26CDB2FD147C880E86A1561F24C1F19F79FB
# gpg: Good signature from "Vladimir Sementsov-Ogievskiy <vsementsov@yandex-team.ru>" [unknown]
# gpg:                 aka "Vladimir Sementsov-Ogievskiy <vsementsov@virtuozzo.com>" [unknown]
# gpg: WARNING: This key is not certified with a trusted signature!
# gpg:          There is no indication that the signature belongs to the owner.
# Primary key fingerprint: 8B9C 26CD B2FD 147C 880E  86A1 561F 24C1 F19F 79FB

* tag 'pull-block-jobs-2026-08-17' of https://gitlab.com/vsementsov/qemu:
  dirty-bitmap: fix integer overflow in serialization coverage
  block/monitor: allow dropping a bitmap never stored on disk
  migration/block-dirty-bitmap: reject bitmap load onto ro node
  block/monitor: reject persistent bitmap add on a read-only node
  tests/unit/test-blockjob: cover keeping a job paused while a pause is pending
  job: keep job paused across overlapping pause requests

+++ block/dirty-bitmap.c
@@ -612,7 +612,7 @@ uint64_t bdrv_dirty_bitmap_serialization_coverage(int serialized_chunk_size,
                                                   const BdrvDirtyBitmap *bitmap)
 {
     uint64_t granularity = bdrv_dirty_bitmap_granularity(bitmap);
-    uint64_t limit = granularity * (serialized_chunk_size << 3);
+    uint64_t limit = granularity * ((uint64_t)serialized_chunk_size << 3);
 
     assert(QEMU_IS_ALIGNED(limit,
                            bdrv_dirty_bitmap_serialization_align(bitmap)));
